Algo-Confusion/requirements.txt
cryptography
pyjwt==1.5.0